changed all async storage to link to firebase

+++ screens/editReminder.js
@@ -1,43 +1,69 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { StatusBar } from "expo-status-bar";
 import { Text, View, StyleSheet, Button, TextInput, Alert } from 'react-native';
 import { ThemedButton } from 'react-native-really-awesome-button';
-import { FIRESTORE_DB } from '../FirebaseConfig';
-import { doc, updateDoc, deleteDoc } from 'firebase/firestore';
+import { FIRESTORE_DB, FIREBASE_AUTH } from '../FirebaseConfig';
+import { doc, updateDoc, deleteDoc, Timestamp } from 'firebase/firestore';
 import DateTimePicker from '@react-native-community/datetimepicker';
+import { onAuthStateChanged } from 'firebase/auth';
 
 export default function EditReminder({ navigation, route }) {
-  const { reminder } = route.params;
+  const { reminder, reminderID } = route.params;
   const [title, setTitle] = useState(reminder.title);
   const [description, setDescription] = useState(reminder.description);
   const [dueDate, setDueDate] = useState(new Date(reminder.dueDate.seconds * 1000));
   const [remind, setRemind] = useState(new Date(reminder.remind.seconds * 1000));
+  const [user, setUser] = useState(null);
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(FIREBASE_AUTH, (user) => {
+        if (user) {
+            setUser(user);
+            console.log('User UID:', user.uid); // Debugging log
+        }
+    });
+
+    return () => unsubscribe();
+  }, []); 
 
   const submitHandler = async () => {
-    const reminderRef = doc(FIRESTORE_DB, 'reminders', reminder.id);
-    await updateDoc(reminderRef, {
-      title: title,
-      description: description,
-      dueDate: dueDate,
-      remind: remind,
-    });
+    if (user) {
+      const uid = user.uid;
+      console.log('Updating reminder for UID:', uid); // Debugging log
+      const reminderRef = doc(FIRESTORE_DB, `users/${uid}/reminders`, reminderID);
+      await updateDoc(reminderRef, {
+        title: title,
+        description: description,
+        dueDate: Timestamp.fromDate(dueDate),
+        remind: Timestamp.fromDate(remind),
+      });
 
-    navigation.navigate('ReminderPage', {
-      reminder: {
-        ...reminder,
-        title,
-        description,
-        dueDate: { seconds: Math.floor(dueDate.getTime() / 1000) },
-        remind: { seconds: Math.floor(remind.getTime() / 1000) },
-      },
-    });
+      navigation.navigate('ReminderPage', {
+        reminder: {
+          ...reminder,
+          title,
+          description,
+          dueDate: { seconds: Math.floor(dueDate.getTime() / 1000) },
+          remind: { seconds: Math.floor(remind.getTime() / 1000) },
+        },
+        reminderID: reminderID
+      });
+    } else {
+      console.error('User is not defined');
+    }
   };
 
   const deleteHandler = async () => {
-    const reminderRef = doc(FIRESTORE_DB, 'reminders', reminder.id);
-    await deleteDoc(reminderRef);
-    navigation.pop();
-    navigation.pop(); 
+    if (user) {
+      const uid = user.uid;
+      console.log('Deleting reminder for UID:', uid); // Debugging log
+      const reminderRef = doc(FIRESTORE_DB, `users/${uid}/reminders`, reminderID);
+      await deleteDoc(reminderRef);
+      navigation.pop();
+      navigation.pop(); 
+    } else {
+      console.error('User is not defined');
+    }
   };
 
   const confirmDelete = () => {
